requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PaperQA2 and its dependencies
paper-qa>=5.0.0
litellm>=1.48.0
langchain>=0.1.0
langchain-community>=0.0.10
pypdf>=3.17.0
pymupdf>=1.23.0
pydantic>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0  # For local embeddings

# Optional but recommended
python-dotenv>=1.0.0
tqdm>=4.66.0